+++ src/main/java/fr/alasdiablo/janoeo/blocks/end/EndRedstoneOre.java
@@ -1,8 +1,13 @@
 package fr.alasdiablo.janoeo.blocks.end;
 
+import fr.alasdiablo.janoeo.blocks.EndOre;
+import net.minecraft.block.BlockState;
 import net.minecraft.block.RedstoneOreBlock;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
 import net.minecraftforge.common.ToolType;
 
 public class EndRedstoneOre extends RedstoneOreBlock {
@@ -16,5 +21,11 @@ public class EndRedstoneOre extends RedstoneOreBlock {
                 .harvestTool(ToolType.PICKAXE)
         );
         this.setRegistryName(registryName);
+    }
+
+    @Override
+    public void onBlockHarvested(World worldIn, BlockPos pos, BlockState state, PlayerEntity player) {
+        super.onBlockHarvested(worldIn, pos, state, player);
+        EndOre.angerEnderman(player, worldIn, pos.getX(), pos.getY(), pos.getZ());
     }
 }